Refactor: back to original links design; italic & skew -> opacity; other styles refactoring
