--- src/components/Header.css
@@ -1,141 +1,0 @@
-.header {
-  max-width: 1900px;
-  width: 100%;
-  margin: 0 auto;
-  display: flex;
-  justify-content: space-between;
-  align-items: flex-start;
-  font-family: 'Inter', Arial, Helvetica, sans-serif;
-  transition: filter .5s ease-in-out;
-  mix-blend-mode: screen;
-  will-change: opacity; /* для корректной работы mix-blend-mode в Safari и на iOS */
-}
-
-.header__marshak-links {
-  margin-left: 2px;
-}
-
-.header__accent-elements-container {
-  margin-bottom: 25px;
-  display: flex;
-  flex-direction: column;
-  align-items: flex-end;
-}
-
-.header__open-live-button {
-  margin-bottom: 8px;
-  border: none;
-  padding: 10px 20px;
-  border-radius: 30px;
-  color: #000;
-  background-color: #fff;
-  cursor: pointer;
-  font-size: 16px;
-  font-weight: 500;
-  line-height: 1;
-  transition: box-shadow .2s ease-out;
-  -webkit-tap-highlight-color: transparent;
-}
-
-.header__open-live-button:focus {
-  outline: none;
-}
-
-.header__open-live-button:hover {
-  box-shadow: 0 0 8px 4px #ffffff40;
-}
-
-.header__form-anchor-link-container {
-  border-radius: 30px;
-  border: 2px solid white;
-  background-color: #fff;
-  transition: box-shadow .2s ease-out;
-  -webkit-tap-highlight-color: transparent;
-}
-
-.header__form-anchor-link-container:hover {
-  box-shadow: 0 0 8px 4px #ffffff40;
-}
-
-.header__form-anchor-link {
-  display: inline-block;
-  padding: 8px 18px;
-  color: #000;
-  font-weight: 500;
-  font-size: 16px;
-  line-height: 1;
-  text-decoration: none;
-  text-align: center;
-  transition: transform .2s ease-out;
-}
-
-.header__form-anchor-link:hover {
-  transform: skewX(-9deg);
-}
-
-@media screen and (max-width: 1024px) {
-  .header__accent-elements-container {
-    margin-bottom: 20px;
-  }
-
-  .header__form-anchor-link {
-    padding: 8px 16px;
-    font-size: 15px;
-  }
-
-  .header__open-live-button {
-    padding: 10px 18px;
-    font-size: 15px;
-  }
-}
-
-@media screen and (max-width: 768px) {
-  .header__marshak-links {
-    margin: 0;
-  }
-}
-
-@media screen and (max-width: 600px) {
-  .header__marshak-links {
-    margin-top: 5px;
-  }
-
-  .header__various-links {
-    margin-top: 5px;
-  }
-}
-
-@media screen and (max-width: 480px) and (min-height: 551px) {
-  .header__accent-elements-container {
-    margin-bottom: 12px;
-  }
-
-  .header__form-anchor-link {
-    padding: 7px 14px;
-    font-size: 12px;
-  }
-
-  .header__open-live-button {
-    padding: 9px 16px;
-    font-size: 12px;
-  }
-}
-
-@media screen and (max-width: 480px) and (max-height: 550px) {
-  .header__accent-elements-container {
-    margin-bottom: 5px;
-    flex-direction: row;
-  }
-
-  .header__form-anchor-link {
-    padding: 7px 14px;
-    font-size: 12px;
-  }
-
-  .header__open-live-button {
-    margin-bottom: 0;
-    margin-right: 5px;
-    padding: 9px 16px;
-    font-size: 12px;
-  }
-}
