+++ src/components/Header/Header.css
@@ -1,0 +1,83 @@
+.header {
+  max-width: 1900px;
+  width: 100%;
+  margin: 0 auto;
+  display: flex;
+  justify-content: space-between;
+  align-items: flex-start;
+  font-family: 'Inter', Arial, Helvetica, sans-serif;
+  transition: filter .5s ease-in-out;
+  mix-blend-mode: screen;
+  will-change: opacity; /* для корректной работы mix-blend-mode в Safari и на iOS */
+}
+
+.header__marshak-links {
+  margin-left: 2px;
+}
+
+.header__accent-elements-container {
+  margin-bottom: 25px;
+  display: flex;
+  flex-direction: column;
+  align-items: flex-end;
+}
+
+.header__open-live-button {
+  margin-bottom: 8px;
+}
+
+@media screen and (max-width: 1024px) {
+  .header__accent-elements-container {
+    margin-bottom: 20px;
+  }
+
+  .header__open-live-button {
+    margin-bottom: 6px;
+  }
+}
+
+@media screen and (max-width: 768px) {
+  .header__marshak-links {
+    margin: 0;
+  }
+}
+
+@media screen and (max-width: 600px) {
+  .header__marshak-links {
+    margin-top: 5px;
+  }
+
+  .header__various-links {
+    margin-top: 5px;
+  }
+}
+
+@media screen and (max-width: 480px) and (min-height: 551px) {
+  .header__accent-elements-container {
+    margin-bottom: 12px;
+  }
+}
+
+@media screen and (max-width: 480px) and (max-height: 550px) {
+  .header__accent-elements-container {
+    margin-bottom: 5px;
+    flex-direction: row;
+  }
+
+  .header__open-live-button {
+    margin-bottom: 0;
+    margin-right: 5px;
+  }
+}
+
+@media screen and (orientation: landscape) and (max-height: 550px) {
+  .header__accent-elements-container {
+    margin-bottom: 5px;
+    flex-direction: row;
+  }
+
+  .header__open-live-button {
+    margin-bottom: 0;
+    margin-right: 5px;
+  }
+}
