+++ src/components/Header/Header.js
@@ -1,12 +1,14 @@
 import './Header.css';
-import './Blur.css';
+import '../Blur.css';
+import '../../style-mixes/main-button/main-button.css';
+import '../../style-mixes/main-link/main-link.css';
 
 import cn from 'classnames';
 import AnchorLink from 'react-anchor-link-smooth-scroll'
 
 import StreamServices from './StreamServices';
 import Marshak from './Marshak';
-import { liveLink } from '../configs/links';
+import { liveLink } from '../../configs/links';
 
 function Header({ isPlayerExtend, isMobile, isLandscape, currentTrack, onVideoModalOpen }) {
   const handleVideoModalOpen = () => {
@@ -29,20 +31,18 @@ function Header({ isPlayerExtend, isMobile, isLandscape, currentTrack, onVideoMo
       <div className="header__various-links">
         <div className="header__accent-elements-container">
           <button
-            className="header__open-live-button"
+            className="main-button main-button_type_accent header__open-live-button"
             onClick={handleVideoModalOpen}
           >
             Live
           </button>
 
-          <div className="header__form-anchor-link-container">
-            <AnchorLink
-              className="header__form-anchor-link"
-              href="#form-participate"
-            >
-              Хочу свой трек
-            </AnchorLink>
-          </div>
+          <AnchorLink
+            className="main-link main-link_type_accent header__form-anchor-link"
+            href="#form-participate"
+          >
+            Хочу свой трек
+          </AnchorLink>
         </div>
 
         <StreamServices
