+++ src/components/Header/StreamServiceLink.css
@@ -1,0 +1,27 @@
+.stream-services-link-container {
+  margin-bottom: 8px;
+}
+
+.stream-services-link-container:last-child {
+  margin-bottom: 0;
+}
+
+
+@media screen and (max-width: 1024px) {
+  .stream-services-link-container {
+    margin-bottom: 6px;
+  }
+}
+
+
+@media screen and (max-width: 480px) and (min-height: 551px) {
+  .stream-services-link-container {
+    margin-bottom: 5px;
+  }
+}
+
+@media screen and (max-width: 480px) and (max-height: 550px) {
+  .stream-services-link-container {
+    margin-bottom: 4px;
+  }
+}
