+++ src/components/Header/StreamServiceLink.js
@@ -1,10 +1,11 @@
 import './StreamServiceLink.css';
+import '../../style-mixes/main-link/main-link.css'
 
 function StreamServiceLink({ serviceLabel, serviceLink, style }) {
   return (
-    <li className="stream-services__link-container" style={style}>
+    <li className="stream-services-link-container" style={style}>
       <a
-        className="stream-services__link"
+        className="main-link main-link_type_common"
         href={serviceLink}
         target="_blank"
         rel="noreferrer"
