--- src/components/MarshakLink.css
@@ -1,61 +1,0 @@
-.marshak__link-container {
-  margin-right: 8px;
-  border-radius: 30px;
-  text-align: center;
-  transition: box-shadow .2s ease-out;
-  -webkit-tap-highlight-color: transparent;
-}
-
-.marshak__link-container:last-child {
-  margin-right: 0px;
-}
-
-.marshak__link-container:hover {
-  box-shadow: 0 0 8px 4px #ffffff40;
-}
-
-.marshak__link {
-  display: inline-block;
-  border-radius: 30px;
-  padding: 8px 18px;
-  color: #fff;
-  text-decoration: none;
-  transition: transform .2s ease-out;
-}
-
-.marshak__link:hover {
-  transform: skewX(-9deg);
-}
-
-@media screen and (max-width: 1024px) {
-  .marshak__link {
-    padding: 8px 16px;
-  }
-}
-
-@media screen and (max-width: 768px) {
-  .marshak__link-container {
-    margin-right: 0;
-    margin-bottom: 8px;
-  }
-}
-
-@media screen and (max-width: 480px) and (min-height: 551px) {
-  .marshak__link-container {
-    margin-bottom: 5px;
-  }
-
-  .marshak__link {
-    padding: 7px 14px;
-  }
-}
-
-@media screen and (max-width: 480px) and (max-height: 550px) {
-  .marshak__link-container {
-    margin-bottom: 4px;
-  }
-
-  .marshak__link {
-    padding: 7px 14px;
-  }
-}
