+++ src/style-mixes/main-button/type/main-button_type_accent.css
@@ -1,0 +1,5 @@
+.main-button_type_accent {
+  color: #000;
+  background-color: #fff;
+  font-weight: 500;
+}
