+++ src/style-mixes/main-button/type/main-button_type_common.css
@@ -1,0 +1,5 @@
+.main-button_type_common {
+  color: #fff;
+  background-color: transparent;
+  font-weight: 400;
+}
