+++ src/style-mixes/main-link/main-link.css
@@ -1,0 +1,38 @@
+@import url(./type/main-link_type_accent.css);
+@import url(./type/main-link_type_common.css);
+
+.main-link {
+  display: inline-block;
+  border: 2px solid white;
+  padding: 8px 18px;
+  border-radius: 30px;
+  font-size: 16px;
+  line-height: 1;
+  text-decoration: none;
+  text-align: center;
+  transition: box-shadow .2s ease-out;
+  -webkit-tap-highlight-color: transparent;
+}
+
+.main-link:hover {
+  box-shadow: 0 0 8px 4px #ffffff40;
+}
+
+.main-link:focus {
+  outline: none;
+  box-shadow: 0 0 8px 4px #ffffff40;
+}
+
+@media screen and (max-width: 1024px) {
+  .main-link {
+    padding: 8px 16px;
+    font-size: 15px;
+  }
+}
+
+@media screen and (max-width: 480px) {
+  .main-link {
+    padding: 7px 14px;
+    font-size: 12px;
+  }
+}
