+++ src/style-mixes/main-link/type/main-link_type_accent.css
@@ -1,0 +1,5 @@
+.main-link_type_accent {
+  background-color: #fff;
+  color: #000;
+  font-weight: 500;
+}
