+++ src/style-mixes/main-link/type/main-link_type_common.css
@@ -1,0 +1,5 @@
+.main-link_type_common {
+  background-color: transparent;
+  color: #fff;
+  font-weight: 400;
+}
